Update Async::Run() and stop using m_scaleMatrix.

+++ libs/glm/detail/qualifier.hpp
@@ -26,7 +26,7 @@ namespace glm
 #		if GLM_CONFIG_ALIGNED_GENTYPES == GLM_ENABLE && defined(GLM_FORCE_DEFAULT_ALIGNED_GENTYPES)
 			defaultp = aligned_lowp
 #		else
-			defaultp = lowp
+			defaultp = highp
 #		endif
 	};
